Start of threadly-test artifact code from 5.42 threadly
Dependencies on threadly core project:
* Clock is depended on most frequently
* ArgumentVerifier (undocumented public API) is used to reduce code-duplication
* NoThreadScheduler depended on by TestableScheduler

The dependency on NoThreadScheduler is the greatest risk of producing byte code incompatibility.
I looked at copying the code and bringing it over, and while possible, it's a very large amount of duplicated logic.  Some of the logic could be simplified (for example NoThreadScheduler extends the task types to allow the time override).  But not very much, it would be ~1k lines of duplicated code to handle the priority based queuing and the NoThreadScheduler concepts.  Similar I am not seeing any impacts to the parent threadly artifacts we could simplify.